we're not allowed to remove internal collections - exclude them from dropping

+++ src/main/java/org/humanbrainproject/knowledgegraph/boundary/indexation/ArangoIndexation.java
@@ -41,7 +41,9 @@ public class ArangoIndexation extends GraphIndexation {
     public void clearGraph() {
         ArangoDatabase db = arango.getOrCreateDB();
         for (CollectionEntity collectionEntity : db.getCollections()) {
-            db.collection(collectionEntity.getName()).drop();
+            if(!collectionEntity.getName().startsWith("_")) {
+                db.collection(collectionEntity.getName()).drop();
+            }
         }
     }
 }
